Problem types and algorithms

Warning

This page is under construction, and may be incomplete.

Depending on the structure a problem can be reduced to, different types of algorithms will apply. The major distinctions are in the number of objective terms, whether any of them is differentiable, whether they are composed with some linear mapping (which in general complicates evaluating the proximal mapping).

Two-terms: $f + g$

This is the most popular model, by far the most thoroughly studied, and an abundance of algorithms exist to solve problems in this form.

Algorithm	Assumptions	Oracle	Implementation	References

Forward-backward	$f$ smooth	$\nabla f$, $\operatorname{prox}_{\gamma g}$	ForwardBackwardIteration	Pierre-Louis Lions, Bertrand Mercier (1979)

Douglas-Rachford		$\operatorname{prox}_{\gamma f}$, $\operatorname{prox}_{\gamma g}$	DouglasRachfordIteration	Jonathan Eckstein, Dimitri P Bertsekas (1992)

Fast forward-backward	$f$ convex, smooth, $g$ convex	$\nabla f$, $\operatorname{prox}_{\gamma g}$	FastForwardBackwardIteration	Paul Tseng (2008), Amir Beck, Marc Teboulle (2009)

PANOC	$f$ smooth	$\nabla f$, $\operatorname{prox}_{\gamma g}$	PANOCIteration	Lorenzo Stella, Andreas Themelis, Pantelis Sopasakis, Panagiotis Patrinos (2017)

ZeroFPR	$f$ smooth	$\nabla f$, $\operatorname{prox}_{\gamma g}$	ZeroFPRIteration	Andreas Themelis, Lorenzo Stella, Panagiotis Patrinos (2018)

Douglas-Rachford line-search	$f$ smooth	$\operatorname{prox}_{\gamma f}$, $\operatorname{prox}_{\gamma g}$	DRLSIteration	Andreas Themelis, Lorenzo Stella, Panagiotis Patrinos (2020)

PANOC+	$f$ locally smooth	$\nabla f$, $\operatorname{prox}_{\gamma g}$	PANOCplusIteration	Alberto De Marchi, Andreas Themelis (2021)

ProximalAlgorithms.ForwardBackwardIteration — Type

ForwardBackwardIteration(; <keyword-arguments>)

Instantiate the forward-backward splitting algorithm (see [1]) for solving optimization problems of the form

minimize f(x) + g(x),

where f is smooth.

Arguments

x0: initial point.

f=Zero(): smooth objective term.

g=Zero(): proximable objective term.

Lf=nothing: Lipschitz constant of the gradient of f.

gamma=nothing: stepsize to use, defaults to 1/Lf if not set (but Lf is).

adaptive=false: forces the method stepsize to be adaptively adjusted.

minimum_gamma=1e-7: lower bound to gamma in case adaptive == true.

References

Lions, Mercier, “Splitting algorithms for the sum of two nonlinear operators,” SIAM Journal on Numerical Analysis, vol. 16, pp. 964–979 (1979).

ProximalAlgorithms.DouglasRachfordIteration — Type

DouglasRachfordIteration(; <keyword-arguments>)

Instantiate the Douglas-Rachford splitting algorithm (see [1]) for solving convex optimization problems of the form

minimize f(x) + g(x).

Arguments

x0: initial point.

f=Zero(): proximable objective term.

g=Zero(): proximable objective term.

gamma: stepsize to use.

References

Eckstein, Bertsekas, "On the Douglas-Rachford Splitting Method and the Proximal Point Algorithm for Maximal Monotone Operators", Mathematical Programming, vol. 55, no. 1, pp. 293-318 (1989).

ProximalAlgorithms.FastForwardBackwardIteration — Type

FastForwardBackwardIteration(; <keyword-arguments>)

Instantiate the accelerated forward-backward splitting algorithm (see [1, 2]) for solving optimization problems of the form

minimize f(x) + g(x),

where f is smooth.

Arguments

x0: initial point.

f=Zero(): smooth objective term.

g=Zero(): proximable objective term.

Lf=nothing: Lipschitz constant of the gradient of f.

gamma=nothing: stepsize to use, defaults to 1/Lf if not set (but Lf is).

adaptive=false: forces the method stepsize to be adaptively adjusted.

minimum_gamma=1e-7: lower bound to gamma in case adaptive == true.

References

Tseng, "On Accelerated Proximal Gradient Methods for Convex-Concave Optimization" (2008).

Beck, Teboulle, "A Fast Iterative Shrinkage-Thresholding Algorithm for Linear Inverse Problems", SIAM Journal on Imaging Sciences, vol. 2, no. 1, pp. 183-202 (2009).

ProximalAlgorithms.PANOCIteration — Type

PANOCIteration(; <keyword-arguments>)

Instantiate the PANOC algorithm (see [1]) for solving optimization problems of the form

minimize f(Ax) + g(x),

where f is smooth and A is a linear mapping (for example, a matrix).

Arguments

x0: initial point.

f=Zero(): smooth objective term.

A=I: linear operator (e.g. a matrix).

g=Zero(): proximable objective term.

Lf=nothing: Lipschitz constant of the gradient of x ↦ f(Ax).

gamma=nothing: stepsize to use, defaults to 1/Lf if not set (but Lf is).

adaptive=false: forces the method stepsize to be adaptively adjusted.

minimum_gamma=1e-7: lower bound to gamma in case adaptive == true.

max_backtracks=20: maximum number of line-search backtracks.

directions=LBFGS(5): strategy to use to compute line-search directions.

References

Stella, Themelis, Sopasakis, Patrinos, "A simple and efficient algorithm for nonlinear model predictive control", 56th IEEE Conference on Decision and Control (2017).

ProximalAlgorithms.ZeroFPRIteration — Type

ZeroFPRIteration(; <keyword-arguments>)

Instantiate the ZeroFPR algorithm (see [1]) for solving optimization problems of the form

minimize f(Ax) + g(x),

where f is smooth and A is a linear mapping (for example, a matrix).

Arguments

x0: initial point.

f=Zero(): smooth objective term.

A=I: linear operator (e.g. a matrix).

g=Zero(): proximable objective term.

Lf=nothing: Lipschitz constant of the gradient of x ↦ f(Ax).

gamma=nothing: stepsize to use, defaults to 1/Lf if not set (but Lf is).

adaptive=false: forces the method stepsize to be adaptively adjusted.

minimum_gamma=1e-7: lower bound to gamma in case adaptive == true.

max_backtracks=20: maximum number of line-search backtracks.

directions=LBFGS(5): strategy to use to compute line-search directions.

References

Themelis, Stella, Patrinos, "Forward-backward envelope for the sum of two nonconvex functions: Further properties and nonmonotone line-search algorithms", SIAM Journal on Optimization, vol. 28, no. 3, pp. 2274–2303 (2018).

ProximalAlgorithms.DRLSIteration — Type

DRLSIteration(; <keyword-arguments>)

Instantiate the Douglas-Rachford line-search algorithm (see [1]) for solving optimization problems of the form

minimize f(x) + g(x),

where f is smooth.

Arguments

x0: initial point.

f=Zero(): smooth objective term.

g=Zero(): proximable objective term.

muf=nothing: convexity modulus of f.

Lf=nothing: Lipschitz constant of the gradient of f.

gamma: stepsize to use, chosen appropriately based on Lf and muf by defaults.

max_backtracks=20: maximum number of line-search backtracks.

directions=LBFGS(5): strategy to use to compute line-search directions.

References

Themelis, Stella, Patrinos, "Douglas-Rachford splitting and ADMM for nonconvex optimization: Accelerated and Newton-type linesearch algorithms", arXiv:2005.10230, 2020.

ProximalAlgorithms.PANOCplusIteration — Type

PANOCplusIteration(; <keyword-arguments>)

Instantiate the PANOCplus algorithm (see [1]) for solving optimization problems of the form

minimize f(Ax) + g(x),

where f is locally smooth and A is a linear mapping (for example, a matrix).

Arguments

x0: initial point.

f=Zero(): smooth objective term.

A=I: linear operator (e.g. a matrix).

g=Zero(): proximable objective term.

Lf=nothing: Lipschitz constant of the gradient of x ↦ f(Ax).

gamma=nothing: stepsize to use, defaults to 1/Lf if not set (but Lf is).

adaptive=false: forces the method stepsize to be adaptively adjusted.

minimum_gamma=1e-7: lower bound to gamma in case adaptive == true.

max_backtracks=20: maximum number of line-search backtracks.

directions=LBFGS(5): strategy to use to compute line-search directions.

References

De Marchi, Themelis, "Proximal gradient algorithms under local Lipschitz gradient continuity: a convergence and robustness analysis of PANOC", arXiv:2112.13000 (2021).

Three-terms: $f + g + h$

When more than one non-differentiable term is there in the objective, algorithms from the previous section do not in general apply out of the box, since $\operatorname{prox}_{\gamma (f + g)}$ does not have a closed form unless in particular cases. Therefore, ad-hoc iteration schemese have been studied.

Algorithm	Assumptions	Oracle	Implementation	References

Davis-Yin	$f, g$ convex, $h$ convex and smooth	$\operatorname{prox}_{\gamma f}$, $\operatorname{prox}_{\gamma g}$, $\nabla h$	DavisYinIteration	Damek Davis, Wotao Yin (2017)

ProximalAlgorithms.DavisYinIteration — Type

DavisYinIteration(; <keyword-arguments>)

Instantiate the Davis-Yin splitting algorithm (see [1]) for solving convex optimization problems of the form

minimize f(x) + g(x) + h(x),

where h is smooth.

Arguments

x0: initial point.

f=Zero(): proximable objective term.

g=Zero(): proximable objective term.

h=Zero(): smooth objective term.

Lh=nothing: Lipschitz constant of the gradient of h.

gamma=nothing: stepsize to use, defaults to 1/Lh if not set (but Lh is).

References

Davis, Yin. "A Three-Operator Splitting Scheme and its Optimization Applications", Set-Valued and Variational Analysis, vol. 25, no. 4, pp. 829–858 (2017).

Primal-dual: $f + g + h \circ L$

When a function $h$ is composed with a linear operator $L$, the proximal operator of $h \circ L$ does not have a closed form in general. For this reason, specific algorithms by the name of "primal-dual" splitting schemes are often applied to this model.

Algorithm	Assumptions	Oracle	Implementation	References

Vu-Condat	$f$ convex and smooth, $g, h$ convex, $L$ linear operator	$\nabla f$, $\operatorname{prox}_{\gamma g}$, $\operatorname{prox}_{\gamma h}$, $L$, $L^*$	VuCodatIteration	Bằng Công Vũ (2013), Laurent Condat (2013)

AFBA	$f$ convex and smooth, $g, h$ convex, $L$ linear operator	$\nabla f$, $\operatorname{prox}_{\gamma g}$, $\operatorname{prox}_{\gamma h}$, $L$, $L^*$	AFBAIteration	Puya Latafat, Panagiotis Patrinos (2017)

ProximalAlgorithms.AFBAIteration — Type

AFBAIteration(; <keyword-arguments>)

Instantiate the asymmetric forward-backward-adjoint algorithm (AFBA, see [1]) for solving convex optimization problems of the form

minimize f(x) + g(x) + (h □ l)(L x),

where f is smooth, g and h are possibly nonsmooth and l is strongly convex. Symbol □ denotes the infimal convolution, and L is a linear mapping.

Points x0 and y0 are the initial primal and dual iterates, respectively. If unspecified, functions f, g, and h default to the identically zero function, l defaults to the indicator of the set {0}, and L defaults to the identity. Important keyword arguments, in case f and l are set, are the Lipschitz constants beta_f and beta_l (see below).

The iterator implements Algorithm 3 of [1] with constant stepsize (α_n=λ) for several prominant special cases:

θ = 2 ==> Corresponds to the Vu-Condat Algorithm [2,3].

θ = 1, μ=1

θ = 0, μ=1

θ ∈ [0,∞), μ=0

See [2, Section 5.2] and [1, Figure 1] for stepsize conditions, special cases, and relation to other algorithms.

Arguments

x0: initial primal point.

y0: initial dual point.

f=Zero(): smooth objective term.

g=Zero(): proximable objective term.

h=Zero(): proximable objective term.

l=IndZero(): strongly convex function.

L=I: linear operator (e.g. a matrix).

beta_f=0: Lipschitz constant of the gradient of f.

beta_l=0: Lipschitz constant of the gradient of l conjugate.

theta=1: nonnegative algorithm parameter.

mu=1: algorithm parameter in the range [0,1].

gamma1: primal stepsize (see [1] for the default choice).

gamma2: dual stepsize (see [1] for the default choice).

References

Latafat, Patrinos, "Asymmetric forward–backward–adjoint splitting for solving monotone inclusions involving three operators", Computational Optimization and Applications, vol. 68, no. 1, pp. 57-93 (2017).

Latafat, Patrinos, "Primal-dual proximal algorithms for structured convex optimization : a unifying framework", In Large-Scale and Distributed Optimization, Giselsson and Rantzer, Eds. Springer International Publishing, pp. 97–120 ( 2018).

Condat, "A primal–dual splitting method for convex optimization involving Lipschitzian, proximable and linear composite terms", Journal of Optimization Theory and Applications, vol. 158, no. 2, pp 460-479 (2013).

Vũ, "A splitting algorithm for dual monotone inclusions involving cocoercive operators", Advances in Computational Mathematics, vol. 38, no. 3, pp. 667-681 (2013).

ProximalAlgorithms.VuCondatIteration — Function

VuCondatIteration(; <keyword-arguments>)

Instantiate the Vũ-Condat primal-dual algorithm (see [1, 2]) for solving convex optimization problems of the form

minimize f(x) + g(x) + (h □ l)(L x),

where f is smooth, g and h are possibly nonsmooth and l is strongly convex. Symbol □ denotes the infimal convolution, and L is a linear mapping.

For the arguments see AFBAIteration.

References

Condat, "A primal–dual splitting method for convex optimization involving Lipschitzian, proximable and linear composite terms", Journal of Optimization Theory and Applications, vol. 158, no. 2, pp 460-479 (2013).

Vũ, "A splitting algorithm for dual monotone inclusions involving cocoercive operators", Advances in Computational Mathematics, vol. 38, no. 3, pp. 667-681 (2013).